fix: enable DCT16/32 transform support in VarDCT encoder
Multiple bugs fixed that prevented DCT16/32 from working:

1. Y plane extraction: compute_ac_strategies() and compute_quant_field()
   now correctly extract Y channel from interleaved XYB data

2. HuffmanNode overflow: Changed index fields from i16 to i32 to support
   alphabets > 32K symbols

3. Quantization formula: Fixed DC/AC coefficient quantization to divide
   by INV_LF_QUANT instead of multiplying (was completely inverting the
   quantization)

4. Canonical Huffman codes: Fixed single-depth prefix code encoding to
   use symbol value ordering (not storage order) when assigning codes

All 407 VarDCT tests now pass including solid colors, gradients,
patterns, and noise with full decode validation.

+++ jxl_enc/src/entropy_coding/huffman_tree.rs
@@ -76,9 +76,10 @@ struct HuffmanNode {
     /// Total count (frequency) of this subtree.
     total_count: u32,
     /// Index of left child, or -1 if leaf.
-    index_left: i16,
+    index_left: i32,
     /// Index of right child, or symbol value if leaf.
-    index_right_or_value: i16,
+    /// Note: libjxl uses i16, but we need i32 for alphabets > 32K symbols.
+    index_right_or_value: i32,
 }
 
 impl HuffmanNode {
@@ -86,7 +87,7 @@ impl HuffmanNode {
         Self {
             total_count: count,
             index_left: -1,
-            index_right_or_value: symbol as i16,
+            index_right_or_value: symbol as i32,
         }
     }
 
@@ -228,8 +229,8 @@ pub fn create_huffman_tree(histogram: &[u32], tree_limit: u8) -> Vec<u8> {
             // Create parent node at the sentinel position
             let j_end = tree.len() - 1;
             tree[j_end].total_count = tree[left].total_count + tree[right].total_count;
-            tree[j_end].index_left = left as i16;
-            tree[j_end].index_right_or_value = right as i16;
+            tree[j_end].index_left = left as i32;
+            tree[j_end].index_right_or_value = right as i32;
 
             // Add new sentinel
             tree.push(HuffmanNode::sentinel());
